+++ examples/calc_metrics.py
@@ -26,7 +26,7 @@ custom metric.
 
 import inspect
 import json
-import traceback
+import logging
 from pathlib import Path
 
 import matplotlib.pyplot as plt
@@ -36,6 +36,8 @@ from l2metrics.core import Metric
 from l2metrics.parser import init_parser
 from l2metrics.report import MetricsReport
 from tqdm import tqdm
+
+logger = logging.getLogger("Calculate Metrics")
 
 
 class MyCustomAgentMetric(Metric):
@@ -97,16 +99,6 @@ def run() -> None:
         ll_metrics_dicts = []
         log_data_df = pd.DataFrame()
 
-        old_print = print
-
-        def new_print(*args, **kwargs):
-            try:
-                tqdm.write(*args, **kwargs)
-            except:
-                old_print(*args, ** kwargs)
-
-        inspect.builtins.print = new_print
-
         # Iterate over all runs found in the directory
         dirs = [p for p in Path(args.log_dir).rglob("*") if p.is_dir()]
         for dir in tqdm(dirs, desc=Path(args.log_dir).name):
@@ -116,8 +108,8 @@ def run() -> None:
                     # Store STE data
                     try:
                         util.store_ste_data(log_dir=dir, mode=args.ste_store_mode)
-                    except Exception as e:
-                        print(e)
+                    except ValueError as e:
+                        logger.error(e)
                 else:
                     # Compute and store the LL metrics
                     kwargs['log_dir'] = dir
@@ -198,8 +190,10 @@ def run() -> None:
                 report.save_settings(output_dir=args.output_dir, filename=args.output)
 
 if __name__ == "__main__":
+    # Configure logger
+    logging.basicConfig(level=logging.INFO)
+
     try:
         run()
-    except Exception as e:
-        print(f'Error: {e}')
-        traceback.print_exc()
+    except (KeyError, ValueError) as e:
+        logger.exception(e)