+++ l2metrics/metrics_json_parser.py
@@ -102,10 +102,6 @@ class MetricsParser:
             if key in a:
                 if isinstance(a[key], dict) and isinstance(b[key], dict):
                     self.mergedict(a[key], b[key], path + [str(key)])
-                # elif a[key] == b[key]:
-                #     pass # same leaf value
-                # else:
-                    # raise Exception('Conflict at %s' % '.'.join(path + [str(key)]))
             else:
                 a[key] = b[key]
         return a