+++ l2metrics/normalizer.py
@@ -19,11 +19,14 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
 IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
+import logging
 from collections import defaultdict
 from typing import Set
 
 import numpy as np
 import pandas as pd
+
+logger = logging.getLogger(__name__)
 
 
 class Normalizer():
@@ -33,7 +36,8 @@ class Normalizer():
     valid_methods = ['task', 'run']
 
     def __init__(self, perf_measure: str, data: pd.DataFrame, ste_data: dict = {},
-                 data_range: defaultdict = None, method: str = 'task', scale: int = 100) -> None:
+                 data_range: defaultdict = None, method: str = 'task', scale: int = 100,
+                 offset: int = 1) -> None:
         """Constructor for Normalizer.
 
         Args:
@@ -46,9 +50,10 @@ class Normalizer():
                 Defaults to 'task'.
             scale (int, optional): Normalization scale, interpreted as from 0 to this value.
                 Defaults to 100.
+            offset (int, optional): Offset to normalized data. Defaults to 1.
 
         Raises:
-            Exception: If input validation fails.
+            ValueError: If data range validation fails.
         """
 
         self.perf_measure = perf_measure
@@ -65,13 +70,16 @@ class Normalizer():
         elif data is not None:
             self.calculate_data_range(data, ste_data)
         else:
-            raise Exception(f'Must provide data or data range to initialize Normalizer')
+            raise ValueError(f'Must provide data or data range to initialize Normalizer')
 
         if self._validate_method(method):
             self.method = method
-        
+
         if self._validate_scale(scale):
             self.scale = scale
+
+        if self._validate_offset(offset):
+            self.offset = offset
 
     def calculate_data_range(self, data: pd.DataFrame, ste_data: dict = {}) -> None:
         """Calculates data range per task for given data.
@@ -84,13 +92,13 @@ class Normalizer():
             ste_data (dict, optional): The STE data for computing quantiles. Defaults to {}.
 
         Raises:
-            Exception: If data contains more than just performance values and task name.
+            ValueError: If data contains more than just performance values and task name.
         """
 
         data_column = data.columns.to_numpy()
 
         if len(data_column) > 1:
-            raise Exception(f'Data must only have one column with performance measures')
+            raise ValueError(f'Data must only have one column with performance measures')
 
         # Initialize data range as empty object
         self.data_range = defaultdict(dict)
@@ -121,20 +129,20 @@ class Normalizer():
             task_names (Set[str]): Set of task names in the data.
 
         Raises:
-            Exception: If data range is not a dictionary object.
-            Exception: If data range is not defined for all tasks.
-            Exception: If the keys min and max are missing.
+            TypeError: If data range is not a dictionary object.
+            KeyError: If data range is not defined for all tasks.
+            KeyError: If the keys min and max are missing.
 
         Returns:
             bool: True if validation succeeds.
         """
 
         if not isinstance(data_range, (dict, defaultdict)):
-            raise Exception(f'Invalid data range type - Must be a dictionary')
+            raise TypeError(f'Invalid data range type - Must be a dictionary')
         elif not set(data_range.keys()).issuperset(task_names):
-            raise Exception(f'Data range not defined for all tasks: {task_names}')
+            raise KeyError(f'Data range not defined for all tasks: {task_names}')
         elif False in [key.keys() >= {'min', 'max'} for key in data_range.values()]:
-            raise Exception(f'Missing required fields: min and max')
+            raise KeyError(f'Missing required fields: min and max')
         else:
             return True
 
@@ -145,14 +153,14 @@ class Normalizer():
             method (str): Normalization method.
 
         Raises:
-            Exception: If method is not in list of valid methods.
+            ValueError: If method is not in list of valid methods.
 
         Returns:
             bool: True if validation succeeds.
         """
 
         if method not in self.valid_methods:
-            raise Exception(f'Invalid normalization method: {method}\n'
+            raise ValueError(f'Invalid normalization method: {method}\n'
                             f'Valid methods are: {self.valid_methods}')
         else:
             return True
@@ -164,17 +172,38 @@ class Normalizer():
             scale (int): Normalization scale.
 
         Raises:
-            Exception: If scale is not an integer.
-            Exception: If scale is less than or equal to 0.
+            TypeError: If scale is not an integer.
+            ValueError: If scale is less than or equal to 0.
 
         Returns:
             bool: True if validation succeeds.
         """
 
         if not isinstance(scale, int):
-            raise Exception(f'Invalid scale type: {type(scale)}')
+            raise TypeError(f'Invalid scale type: {type(scale)}')
         elif scale <= 0:
-            raise Exception(f'Scale value must be greater than 0: {scale}')
+            raise ValueError(f'Scale value must be greater than 0: {scale}')
+        else:
+            return True
+
+    def _validate_offset(self, offset: int) -> bool:
+        """Validates normalization offset.
+
+        Args:
+            offset (int): Normalization offset.
+
+        Raises:
+            TypeError: If offset is not an integer.
+            ValueError: If scale is less than or equal to 0.
+
+        Returns:
+            bool: True if validation succeeds.
+        """
+
+        if not isinstance(offset, int):
+            raise TypeError(f'Invalid offset type: {type(offset)}')
+        elif offset <= 0:
+            raise ValueError(f'Offset value must be greater than 0: {offset}')
         else:
             return True
 
@@ -182,7 +211,10 @@ class Normalizer():
         """Normalizes the given data with the current instance method and data range/scale. 
 
         Args:
-            data (pd.DataFrame): Dataframe to be normalized
+            data (pd.DataFrame): Dataframe to be normalized.
+
+        Raises:
+            KeyError: If there's a missing data range for any task.
 
         Returns:
             pd.DataFrame: Normalized dataframe.
@@ -193,18 +225,20 @@ class Normalizer():
                 if task in self.data_range.keys():
                     task_min = self.data_range[task]['min']
                     task_max = self.data_range[task]['max']
+                    task_data = data.loc[data['task_name'] == task, self.perf_measure].to_numpy()
 
                     if task_min == task_max:
-                        raise Exception(f"Cannot normalize performance for task ({task}) - " \
-                            "Performance is constant")
-
-                    data.loc[data.task_name == task, self.perf_measure] = \
-                        ((data[data['task_name'] == task][self.perf_measure].to_numpy() - task_min) / \
-                        (task_max - task_min) * self.scale) + 1
+                        data.loc[data['task_name'] == task, self.perf_measure] = (
+                            task_data * 0) + self.offset
+                        logger.warning(f"Performance for task ({task}) is constant - normalizing to 0")
+                    else:
+                        data.loc[data['task_name'] == task, self.perf_measure] = \
+                            ((task_data - task_min) / (task_max - task_min)
+                             * self.scale) + self.offset
                 else:
-                    raise Exception(f"Missing data range for task '{task}'")
+                    raise KeyError(f"Missing data range for task '{task}'")
             return data
         elif self.method == 'run':
-            data.loc[:, self.perf_measure] = ((data[self.perf_measure].to_numpy() - self.run_min) / \
-                (self.run_max - self.run_min) * self.scale) + 1
+            data.loc[:, self.perf_measure] = ((data[self.perf_measure].to_numpy() - self.run_min) /
+                                              (self.run_max - self.run_min) * self.scale) + self.offset
             return data